+++ reviver/lib/ctors.js
@@ -22,6 +22,7 @@
 
 var Float64Array = require( './../../float64' );
 var Float32Array = require( './../../float32' );
+var Float16Array = require( './../../float16' );
 var Int16Array = require( './../../int16' );
 var Int32Array = require( './../../int32' );
 var Int8Array = require( './../../int8' );
@@ -39,6 +40,7 @@ var BooleanArray = require( './../../bool' );
 var ctors = {
 	'Float64Array': Float64Array,
 	'Float32Array': Float32Array,
+	'Float16Array': Float16Array,
 	'Int32Array': Int32Array,
 	'Uint32Array': Uint32Array,
 	'Int16Array': Int16Array,